Add global ccr stats endpoint and tidy up implemented ccr api's thus far
