+++ src/Nest/XPack/CrossClusterReplication/Stats/CcrStatsRequest.cs
@@ -3,30 +3,18 @@ using Newtonsoft.Json;
 
 namespace Nest
 {
+	/// <summary>
+	/// This API gets cross-cluster replication stats. This API will return all stats related to cross-cluster replication.
+	/// In particular, this API returns stats about auto-following, and returns the same shard-level stats as in the get
+	/// follower stats API. <see cref="IElasticClient.FollowIndexStats(Nest.Indices,System.Func{Nest.FollowIndexStatsDescriptor,Nest.IFollowIndexStatsRequest})"/>
+	/// </summary>
 	[MapsApi("ccr.stats.json")]
 	[ContractJsonConverter(typeof(ReadAsTypeJsonConverter<CcrStatsRequest>))]
-	public partial interface ICcrStatsRequest
-	{
-		/// <summary>
-		///
-		/// <para>
-		/// </para>
-		/// </summary>
-		[JsonProperty("cursor")]
-		string Cursor { get; set; }
-	}
+	public partial interface ICcrStatsRequest { }
 
-	public partial class CcrStatsRequest
-	{
-		/// <inheritdoc cref="ICcrStatsRequest.Cursor" />
-		public string Cursor { get; set; }
-	}
+	/// <inheritdoc cref="ICcrStatsRequest"/>
+	public partial class CcrStatsRequest { }
 
-	public partial class CcrStatsDescriptor
-	{
-		string ICcrStatsRequest.Cursor { get; set; }
-
-		/// <inheritdoc cref="ICcrStatsRequest.Cursor" />
-		public CcrStatsDescriptor Cursor(string cursor) => Assign(a => a.Cursor = cursor);
-	}
+	/// <inheritdoc cref="ICcrStatsRequest"/>
+	public partial class CcrStatsDescriptor { }
 }
