+++ src/Nest/XPack/CrossClusterReplication/Stats/ElasticClient-CcrStats.cs
@@ -7,7 +7,11 @@ namespace Nest
 {
 	public partial interface IElasticClient
 	{
-		/// <summary></summary>
+		/// <summary>
+		/// This API gets cross-cluster replication stats. This API will return all stats related to cross-cluster replication.
+		/// In particular, this API returns stats about auto-following, and returns the same shard-level stats as in the get
+		/// follower stats API. <see cref="IElasticClient.FollowIndexStats(Nest.Indices,System.Func{Nest.FollowIndexStatsDescriptor,Nest.IFollowIndexStatsRequest})"/>
+		/// </summary>
 		ICcrStatsResponse CcrStats(Func<CcrStatsDescriptor, ICcrStatsRequest> selector = null);
 
 		/// <inheritdoc cref="CcrStats(System.Func{Nest.CcrStatsDescriptor,Nest.ICcrStatsRequest})" />
